+++ src/experimental/cellular-automaton/scenes/const/const.ts
@@ -1,0 +1,8 @@
+export const CONST = {
+  CELLSIZE: 6,
+  GRID_WIDTH: 100,
+  GRID_HEIGHT: 100,
+  OFFSET: 0,
+  FIELD_COLOR: 0xffffff,
+  FIELD_ALPHA: 1,
+};