+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Add new `experimental` project `lissajous-curve` *[Author: digitsensitive (2020-03-21)]*
 * Add new game `candy-crush` *[Author: digitsensitive (2020-03-26)]*
 * Add new `experimental` project `procedural-generation` *[Author: digitsensitive (2020-04-12)]*
+* Add new `experimental` project `cellular-automaton` *[Author: digitsensitive (2020-05-05)]*
 
 ### Updates